Issue #117: Update search-updating logic
The Search_Job logic gets invoked in 4 contexts:

- To generate results for an Exporter
- To generate push-based alerts
- An ad-hoc execution
- A scheduled execution

For the first 3 scenarios, we don't want to update any of the
bookkeeping that the Search does. The final scenario is a bit more
nuanced.

We can further break this down into 3 sub-scenarios:

- A new job is executed
- A previously failed job is re-executed
- A backfilled job is executed

Previously, all three of these sub-scenarios would trigger emails and
update the associated Search. However, we should only be running this
update logic in the first case - the other 2 scenarios are running over
a time-window in the past.